requirements.txt
# Core dependencies
flask==2.3.3
python-dotenv==1.0.0
pymongo==4.5.0
requests==2.31.0
beautifulsoup4==4.12.2

# LLM providers (uncomment based on your choice)
# openai==1.3.0
# anthropic==0.8.0

# Optional - for vector search capabilities
# langchain==0.0.292
# langchain-community==0.0.13
# langchain-openai==0.0.2
# numpy==1.24.4